Move note calculation logic from Keyboard to NoteCalculator
