+++ keyboard.hpp
@@ -1,17 +1,11 @@
 #pragma once
 
-#include "notes.hpp"
+#include "note.hpp"
+#include "note-calculator.hpp"
 #include "octave.hpp"
 
 class Keyboard {
 public:
-  // Resistance of all resistors in the keyboard circuit.
-  // R1 should be equal 10kOm.
-  // Resistance of R2-R8 should be sorted ascending.
-  struct Resistors {
-    unsigned long R1, R2, R3, R4, R5, R6, R7, R8;
-  };
-
   Keyboard(Resistors,
            uint8_t analogPin,
            uint8_t decreaseOctaveButtonPin,
@@ -28,22 +22,8 @@ public:
   [[nodiscard]] int GetOctave() const;
 
 private:
-  [[nodiscard]] int calculateRealVoltage(int adcVoltage);
-  [[nodiscard]] Note matchVoltageWithNote(int voltage) const;
-
-  [[nodiscard]] int calculateNoteVoltage(unsigned long noteResistance) const;
-
-  const Resistors RESISTORS;
-  const Note NOTES[NOTES_AMOUNT]{ Note::C4, Note::C4_SHARP,
-                                  Note::D4, Note::D4_SHARP,
-                                  Note::E4,
-                                  Note::F4, Note::F4_SHARP,
-                                  Note::G4, Note::G4_SHARP,
-                                  Note::A4, Note::A4_SHARP,
-                                  Note::B4 };
-  const int NOTES_VOLTAGE[NOTES_AMOUNT];
-
   const uint8_t ANALOG_PIN;
-  Note m_CurrentNote{};
+  Note m_CurrentNote{ Note::NONE };
+  const NoteCalculator NOTE_CALCULATOR;
   Octave m_Octave;
 };
